corrected typo in fft functions

+++ tfkbnufft/nufft/interp_functions.py
@@ -72,7 +72,7 @@ def calc_coef_and_indices(tm, kofflist, Jval, table, centers, L, dims, conjcoef=
     gridind = tf.cast(gridind, int_type)
 
     arr_ind = tf.zeros((M,), dtype=int_type)
-    coef = tf.ones(M, dtype=table.dtype)
+    coef = tf.ones(M, dtype=table[0].dtype)
 
     for d in range(ndims):  # spatial dimension
         sliced_table = tf.gather_nd(table[d], (distind[d, :] + centers[d])[:, None])
